Work towards running the simulation in batches of time
Goal to be able to run the SUMO simulation for longer durations and then run IoTOsmosisRES in batches, i.e. for SUMO time 0-3600 the IoT simulation could be run 0-200 and then 200-400 etc. Current implementation still needs to account for communications which start in one batch but then finish in next batch