Changes made to allow for the simulation of an external simulator injecting events into SimulatorBridger. And changes made to cellFree for MCFR
To properly use with an external simualtor the current simualtion time is needed, and so sb.run() now returns this time, as each wakeUp time is not necessarily +deltaTime after the last wakeUptime. Changes were made to the cellFree algorithm as MCFR needs earlier access to which MEL is being actually used, the previous cellFree was only tested with SPMB but it changes the chosen MEL after the MCFR had routed for 'old' MEL which caused issues/crashes. Currently injected devices to not deplete battery. Changes made to fix issues with injected devices not properly tracking how many packets were being sent for accurateBatteryInfo.csv Also some changes made to improve performance.